styles and hash not working
