+++ js/app/components/IFrame/iframe-css.css
@@ -3,3 +3,8 @@
 	width: 100%;
 	height: 50vh;
 }
+
+:local(.appFrame) {
+	border: 0;
+	width: 100%;
+}
