+++ js/app/components/InvoiceLogin/login-page.css
@@ -3,7 +3,6 @@
 }
 
 :local(.invoicesLogin) {
-	height: 70vh;
 	display: flex;
 	justify-content: center;
 	align-items: center;
